[examples] Moves nlp examples into nlp folder
