+++ examples/docs/sentiment_analysis.md
@@ -2,7 +2,7 @@
 
 In this example, you learn how to use the DistilBERT model trained by HuggingFace using PyTorch. 
 You can provide the model with a question and a paragraph containing an answer. The model is then able to find the best answer from the answer paragraph.
-You can find the source code in [SentimentAnalysis.java](https://github.com/deepjavalibrary/djl/blob/master/examples/src/main/java/ai/djl/examples/inference/SentimentAnalysis.java).
+You can find the source code in [SentimentAnalysis.java](https://github.com/deepjavalibrary/djl/blob/master/examples/src/main/java/ai/djl/examples/inference/nlp/SentimentAnalysis.java).
 
 Example:
 
@@ -29,5 +29,5 @@ Follow [setup](../../docs/development/setup.md) to configure your development en
 
 ```sh
 cd examples
-./gradlew run -Dmain=ai.djl.examples.inference.SentimentAnalysis
+./gradlew run -Dmain=ai.djl.examples.inference.nlp.SentimentAnalysis
 ```
