+++ examples/src/main/java/ai/djl/examples/inference/nlp/BertClassification.java
@@ -10,7 +10,7 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package ai.djl.examples.inference;
+package ai.djl.examples.inference.nlp;
 
 import ai.djl.MalformedModelException;
 import ai.djl.ModelException;
