+++ examples/src/test/java/ai/djl/examples/inference/cv/BigGANTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
  * with the License. A copy of the License is located at
@@ -10,11 +10,11 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package ai.djl.examples.inference;
+package ai.djl.examples.inference.cv;
 
 import ai.djl.ModelException;
-import ai.djl.examples.inference.cv.PoseEstimation;
-import ai.djl.modality.cv.output.Joints;
+import ai.djl.modality.cv.Image;
+import ai.djl.testing.TestRequirements;
 import ai.djl.translate.TranslateException;
 
 import org.testng.Assert;
@@ -22,12 +22,18 @@ import org.testng.annotations.Test;
 
 import java.io.IOException;
 
-public class PoseEstimationTest {
+public class BigGANTest {
 
     @Test
-    public void testPoseEstimation() throws ModelException, TranslateException, IOException {
-        Joints[] result = PoseEstimation.predict();
-        Assert.assertEquals(result.length, 3);
-        Assert.assertEquals(result[0].getJoints().size(), 17);
+    public void testBigGAN() throws ModelException, TranslateException, IOException {
+        TestRequirements.linux();
+
+        Image[] generatedImages = BigGAN.generate();
+
+        Assert.assertEquals(generatedImages.length, 5);
+        for (Image img : generatedImages) {
+            Assert.assertEquals(img.getWidth(), 256);
+            Assert.assertEquals(img.getHeight(), 256);
+        }
     }
 }
