+++ examples/src/test/java/ai/djl/examples/inference/cv/InstanceSegmentationTest.java
@@ -10,34 +10,25 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package ai.djl.examples.inference;
+package ai.djl.examples.inference.cv;
 
 import ai.djl.ModelException;
-import ai.djl.examples.inference.face.RetinaFaceDetection;
 import ai.djl.modality.Classifications;
 import ai.djl.modality.cv.output.DetectedObjects;
-import ai.djl.testing.TestRequirements;
 import ai.djl.translate.TranslateException;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
 import java.io.IOException;
-import java.util.Collections;
-import java.util.List;
 
-public class RetinaFaceDetectionTest {
+public class InstanceSegmentationTest {
 
     @Test
-    public void testRetinaFaceDetection() throws ModelException, TranslateException, IOException {
-        TestRequirements.linux();
-        TestRequirements.nightly();
-        DetectedObjects result = RetinaFaceDetection.predict();
-
-        List<String> objects = Collections.singletonList("Face");
-        for (Classifications.Classification obj : result.items()) {
-            Assert.assertTrue(objects.contains(obj.getClassName()));
-            Assert.assertTrue(Double.compare(obj.getProbability(), 0.6) > 0);
-        }
+    public void testInstanceSegmentation() throws ModelException, TranslateException, IOException {
+        DetectedObjects result = InstanceSegmentation.predict();
+        Classifications.Classification best = result.best();
+        Assert.assertEquals(best.getClassName(), "bicycle");
+        Assert.assertTrue(Double.compare(best.getProbability(), 0.7) > 0);
     }
 }
