+++ examples/src/test/java/ai/djl/examples/inference/cv/ObjectDetectionWithTensorflowSavedModelTest.java
@@ -10,10 +10,10 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package ai.djl.examples.inference;
+
+package ai.djl.examples.inference.cv;
 
 import ai.djl.ModelException;
-import ai.djl.examples.inference.face.LightFaceDetection;
 import ai.djl.modality.Classifications;
 import ai.djl.modality.cv.output.DetectedObjects;
 import ai.djl.testing.TestRequirements;
@@ -23,21 +23,26 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 
 import java.io.IOException;
-import java.util.Collections;
+import java.util.Arrays;
 import java.util.List;
 
-public class LightFaceDetectionTest {
+public class ObjectDetectionWithTensorflowSavedModelTest {
 
     @Test
-    public void testLightFaceDetection() throws ModelException, TranslateException, IOException {
+    public void testObjectDetection() throws ModelException, TranslateException, IOException {
+        // Only run nightly, this example download the synset file from github, this can cause
+        // throttling and will fail the test.
         TestRequirements.linux();
+        TestRequirements.nightly();
+        TestRequirements.notGpu();
 
-        DetectedObjects result = LightFaceDetection.predict();
+        DetectedObjects result = ObjectDetectionWithTensorflowSavedModel.predict();
 
-        List<String> objects = Collections.singletonList("Face");
+        Assert.assertEquals(result.getNumberOfObjects(), 3);
+        List<String> objects = Arrays.asList("dog", "bicycle", "car");
         for (Classifications.Classification obj : result.items()) {
             Assert.assertTrue(objects.contains(obj.getClassName()));
-            Assert.assertTrue(Double.compare(obj.getProbability(), 0.6) > 0);
+            Assert.assertTrue(Double.compare(obj.getProbability(), 0.7) > 0);
         }
     }
 }
