+++ examples/src/test/java/ai/djl/examples/inference/face/FeatureComparisonTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
  * with the License. A copy of the License is located at
@@ -10,10 +10,9 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package ai.djl.examples.inference;
+package ai.djl.examples.inference.face;
 
 import ai.djl.ModelException;
-import ai.djl.examples.inference.cv.StyleTransfer;
 import ai.djl.modality.cv.Image;
 import ai.djl.modality.cv.ImageFactory;
 import ai.djl.testing.TestRequirements;
@@ -23,19 +22,26 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 
 import java.io.IOException;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 
-public class StyleTransferTest {
+public class FeatureComparisonTest {
 
     @Test
-    public void testStyleTransfer() throws IOException, ModelException, TranslateException {
+    public void testFeatureComparison() throws ModelException, TranslateException, IOException {
         TestRequirements.linux();
 
-        String imagePath = "src/test/resources/mountains.png";
-        Image input = ImageFactory.getInstance().fromFile(Paths.get(imagePath));
-        Image generatedImage = StyleTransfer.transfer(input, StyleTransfer.Artist.MONET);
+        TestRequirements.nightly();
 
-        Assert.assertEquals(generatedImage.getWidth(), input.getWidth());
-        Assert.assertEquals(generatedImage.getHeight(), input.getHeight());
+        Path imageFile1 = Paths.get("src/test/resources/kana1.jpg");
+        Image img1 = ImageFactory.getInstance().fromFile(imageFile1);
+        Path imageFile2 = Paths.get("src/test/resources/kana2.jpg");
+        Image img2 = ImageFactory.getInstance().fromFile(imageFile2);
+
+        float[] feature1 = FeatureExtraction.predict(img1);
+        float[] feature2 = FeatureExtraction.predict(img2);
+
+        Assert.assertTrue(
+                Double.compare(FeatureComparison.calculSimilar(feature1, feature2), 0.6) > 0);
     }
 }
